maint: Initial implementation of lp_scene_t
Scenes (lp_scene_t objects) are implemented as subclass of GObject. We are
going to use the GLib type system to implement Scenes and Media objects. This
is the first commit towards this goal. We are going to keep this changes on
the branch play-gobject until we have a stable implementation.

  * lib/Makefile.am
  * lib/lp-scene.c
  * lib/play.h