maint: Add support to lp_Media selection
When one clicks on an lp_Media, besides the regular lp_EventPointerClick
event created having a lp_Scene as source, another event of the same type
is created having the corresponding lp_Media as source. In this case, the "x"
and "y" properties of the event are relative to the "x" and "y" properties of
the lp_Media.

Thus, an lp_Event of mask LP_EVENT_MASK_POINTER_CLICK having as source an
lp_Media represents a click over that media (e.g., a selection).

* lib/lp-event-pointer-click.c
* lib/lp-media.c
* lib/lp-scene.c
* lib/play-internal.h