maint: Fix bug when pausing a lp_Scene
When calling `lp_scene_pause`, the LP_EVENT's source is a lp_Scene and not
a lp_Media. We had a cast to LP_MEDIA to whatever event source received
from the following masks:
  * LP_EVENT_MASK_ERROR
  * LP_EVENT_MASK_START
  * LP_EVENT_MASK_STOP
  * LP_EVENT_MASK_PAUSE
  * LP_EVENT_MASK_SEEK

Now we first check if the source is indeed a lp_Media before trying to apply
the cast (all this occurs in the `lp_scene_bus_callback` -- lp-scene.c).

* lib/lp-scene.c